# shared/contracts/openapi/ai-mode.v1.openapi.json
{
  "components": {
    "parameters": {
      "IdempotencyKey": {
        "in": "header",
        "name": "Idempotency-Key",
        "required": false,
        "schema": {
          "maxLength": 200,
          "minLength": 1,
          "type": "string"
        }
      },
      "RequestId": {
        "in": "header",
        "name": "X-Request-ID",
        "required": false,
        "schema": {
          "maxLength": 200,
          "type": "string"
        }
      },
      "Traceparent": {
        "in": "header",
        "name": "traceparent",
        "required": false,
        "schema": {
          "pattern": "^00-[0-9a-f]{32}-[0-9a-f]{16}-[0-9a-f]{2}$",
          "type": "string"
        }
      }
    },
    "schemas": {
...
      "AgentRun": {
        "additionalProperties": false,
        "description": "Public, safe snapshot of a persisted agent run.",
        "properties": {
          "cancel_requested": {
            "default": false,
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolError"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "feature_key": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9_.-]*$",
            "title": "Feature Key",
            "type": "string"
          },
          "final_result": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Final Result"
          },
          "grounding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroundingRequest"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "iteration_count": {
            "default": 0,
            "minimum": 0,
            "title": "Iteration Count",
            "type": "integer"
          },
          "limits": {
            "$ref": "#/components/schemas/RunLimits"
          },
          "model_profile": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9_.-]*$",
            "title": "Model Profile",
            "type": "string"
          },
          "objective": {
            "maxLength": 16000,
            "minLength": 1,
            "title": "Objective",
            "type": "string"
          },
          "prompt_set": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9_.-]*$",
            "title": "Prompt Set",
            "type": "string"
          },
          "request_id": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:/-]{0,199}$",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "tool_allowlist": {
            "anyOf": [
              {
                "items": {
                  "maxLength": 100,
                  "minLength": 1,
                  "pattern": "^[a-z0-9][a-z0-9_.-]*$",
                  "type": "string"
                },
                "maxItems": 50,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Tool Allowlist"
          },
          "tool_call_count": {
            "default": 0,
            "minimum": 0,
            "title": "Tool Call Count",
            "type": "integer"
          },
          "traceparent": {
            "anyOf": [
              {
                "pattern": "^00-[0-9a-f]{32}-[0-9a-f]{16}-[0-9a-f]{2}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Traceparent"
          },
          "trusted_identifiers": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TrustedIdentifier"
            },
            "maxItems": 100,
            "title": "Trusted Identifiers",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "default": 0,
            "minimum": 0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "request_id",
          "feature_key",
          "objective",
          "status",
          "prompt_set",
          "model_profile",
          "limits",
          "created_at",
          "updated_at"
        ],
        "title": "AgentRun",
        "type": "object"
      },
      "AgentRunDetail": {
        "additionalProperties": false,
        "description": "Run snapshot plus its ordered, safe phase history.",
        "properties": {
          "reviews": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/HumanReview"
            },
            "title": "Reviews",
            "type": "array"
          },
          "run": {
            "$ref": "#/components/schemas/AgentRun"
          },
          "steps": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AgentStep"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "run"
        ],
        "title": "AgentRunDetail",
        "type": "object"
      },
...
      "AgentRunEventPage": {
        "additionalProperties": false,
        "description": "Bounded cursor page of safe progress events.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AgentRunEvent"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "minimum": 0,
            "title": "Next Cursor",
            "type": "integer"
          },
          "terminal": {
            "title": "Terminal",
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "next_cursor",
          "terminal"
        ],
        "title": "AgentRunEventPage",
        "type": "object"
      },
...
      "AgentRunPage": {
        "additionalProperties": false,
        "description": "Stable cursor page of operations summaries.",
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "items": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AgentRunSummary"
            },
            "maxItems": 100,
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Next Cursor"
          }
        },
        "required": [
          "as_of"
        ],
        "title": "AgentRunPage",
        "type": "object"
      },
      "AgentRunRequest": {
        "additionalProperties": false,
        "description": "Request accepted from a feature backend to start one agent run.",
        "properties": {
          "feature_key": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9_.-]*$",
            "title": "Feature Key",
            "type": "string"
          },
          "grounding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroundingRequest"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "limits": {
            "$ref": "#/components/schemas/RunLimits"
          },
          "model_profile": {
            "default": "remote-standard.v1",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9_.-]*$",
            "title": "Model Profile",
            "type": "string"
          },
          "objective": {
            "maxLength": 16000,
            "minLength": 1,
            "title": "Objective",
            "type": "string"
          },
          "prompt_set": {
            "default": "default.v7",
            "enum": [
              "default.v1",
              "default.v2",
              "default.v3",
              "default.v4",
              "default.v5",
              "default.v6",
              "default.v7",
              "default.v8"
            ],
            "title": "Prompt Set",
            "type": "string"
          },
          "tool_allowlist": {
            "anyOf": [
              {
                "items": {
                  "maxLength": 100,
                  "minLength": 1,
                  "pattern": "^[a-z0-9][a-z0-9_.-]*$",
                  "type": "string"
                },
                "maxItems": 50,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Tool Allowlist"
          },
          "trusted_identifiers": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TrustedIdentifier"
            },
            "maxItems": 100,
            "title": "Trusted Identifiers",
            "type": "array"
          }
        },
        "required": [
          "feature_key",
          "objective"
        ],
        "title": "AgentRunRequest",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "ProblemDetail": {
        "additionalProperties": false,
        "description": "Problem Details-compatible error response with stable project extensions.",
        "properties": {
          "code": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "errors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FieldIssue"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Request Id"
          },
          "status": {
            "maximum": 599,
            "minimum": 400,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "RunStatus": {
        "description": "Persisted lifecycle of an agent run.",
        "enum": [
          "queued",
          "planning",
          "ready",
          "acting",
          "observing",
          "adapting",
          "review_required",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Release 0 shared bounded agent-run orchestration API.",
    "title": "ASD AI-mode API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/agent-runs": {
      "get": {
        "operationId": "listAgentRuns",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "items": {
                "$ref": "#/components/schemas/RunStatus"
              },
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "feature_key",
            "required": false,
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "model_profile",
            "required": false,
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "maxLength": 512,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunPage"
                }
              }
            },
            "description": "Stable page of safe agent-run summaries"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem Details-compatible error"
          }
        }
      },
      "post": {
        "operationId": "createAgentRun",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRun"
                }
              }
            },
            "description": "Run persisted and accepted",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Agent-Run-ID": {
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem Details-compatible error"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem Details-compatible error"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem Details-compatible error"
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem Details-compatible error"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem Details-compatible error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem Details-compatible error"
          }
        }
      }
    },
    "/api/v1/agent-runs/{run_id}": {
      "get": {
        "operationId": "getAgentRun",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunDetail"
                }
              }
            },
            "description": "Current run and safe ordered history"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem Details-compatible error"
          }
        }
      }
    },
    "/api/v1/agent-runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancelAgentRun",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRun"
                }
              }
            },
            "description": "Cancellation intent recorded"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem Details-compatible error"
          }
        }
      }
    },
    "/api/v1/agent-runs/{run_id}/events": {
      "get": {
        "operationId": "getAgentRunEvents",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunEventPage"
                }
              }
            },
            "description": "Ordered safe progress events after the cursor"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem Details-compatible error"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem Details-compatible error"
          }
        }
      }
    },
    "/api/v1/agent-runs/{run_id}/reviews": {
      "post": {
        "operationId": "reviewAgentRun",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunDetail"
                }
              }
            },
            "description": "Review atomically applied"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem Details-compatible error"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem Details-compatible error"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem Details-compatible error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem Details-compatible error"
          }
        }
      }
    },
    "/api/v1/model-profiles": {
      "get": {
        "operationId": "getModelProfiles",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelRegistry"
                }
              }
            },
            "description": "Supported models and bounded runtime profiles"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Problem Details-compatible error"
          }
        }
      }
    },
...
  }
}